requirements.txt
streamlit
google-generativeai
openai
numpy
scikit-learn
pydantic<3
PyPDF2
pdfminer.six
reportlab
faster-whisper
pydub
PyYAML
tiktoken
streamlit-webrtc
python-dotenv
